+++ clang-tools-extra/clang-tidy/plex/AvoidStdCallOnceCheck.h
@@ -1,0 +1,32 @@
+//===--- AvoidStdCallOnceCheck.h - clang-tidy -------------------*- C++ -*-===//
+//
+//                     The LLVM Compiler Infrastructure
+//
+// This file is distributed under the University of Illinois Open Source
+// License. See LICENSE.TXT for details.
+//
+//===----------------------------------------------------------------------===//
+
+#ifndef LLVM_CLANG_TOOLS_EXTRA_CLANG_TIDY_PLEX_AVOIDSTDCALLONCECHECK_H
+#define LLVM_CLANG_TOOLS_EXTRA_CLANG_TIDY_PLEX_AVOIDSTDCALLONCECHECK_H
+
+#include "../ClangTidy.h"
+
+namespace clang {
+namespace tidy {
+namespace plex {
+
+/// Disallow std::call_once.
+class AvoidStdCallOnceCheck : public ClangTidyCheck {
+public:
+  AvoidStdCallOnceCheck(StringRef Name, ClangTidyContext *Context)
+      : ClangTidyCheck(Name, Context) {}
+  void registerMatchers(ast_matchers::MatchFinder *Finder) override;
+  void check(const ast_matchers::MatchFinder::MatchResult &Result) override;
+};
+
+} // namespace plex
+} // namespace tidy
+} // namespace clang
+
+#endif // LLVM_CLANG_TOOLS_EXTRA_CLANG_TIDY_PLEX_AVOIDSTDCALLONCECHECK_H